agents/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# LangChain and LangGraph
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20

# HTTP client
httpx==0.25.2

# Utilities
python-multipart==0.0.6
